+++ src/components/Dropdown/types/dropdownLinks.ts
@@ -1,0 +1,5 @@
+export type dropdownLinksType = {
+    id: number;
+    name: string;
+    link: string;
+}
